lib: fwts_dump_data: tidy up code, ensure we limit nbytes to 16
Tidy up this a little.  Make buffer sizes type size_t. Make sure
that we can only dump a maximum of 16 bytes.  Tidy up the formatting
of the function declaration.

Acked-by: Keng-Yu Lin <kengyu@canonical.com>
Acked-by: Alex Hung <alex.hung@canonical.com>

+++ src/lib/include/fwts_dump_data.h
@@ -22,6 +22,6 @@
 
 #include "fwts.h"
 
-void fwts_dump_raw_data(char *buffer, const int len, const uint8_t *data, const int where, const int bytes);
+void fwts_dump_raw_data(char *buffer, const size_t len, const uint8_t *data, const int where, const size_t bytes);
 
 #endif
